+++ docs/release-notes.md
@@ -4,6 +4,19 @@
 ![PowerShell Gallery](https://img.shields.io/powershellgallery/dt/Scriptbook.svg?label=PSGallery%20Downloads&logo=PowerShell&style=flat-square)
 [![Build Status](https://dev.azure.com/tedon/TD.Deploy/_apis/build/status/ehagen.Scriptbook?branchName=master)](https://dev.azure.com/tedon/TD.Deploy/_build/latest?definitionId=52&branchName=master)
 
+## [0.5.5]
+
+```plain
+### Added
+
+- Added -Confirm support to Start-Workflow
+
+### Changed
+
+- Fixed -WhatIf precedence over -Confirm like PowerShell does in 'SupportsShouldProcess'
+
+```
+
 ## [0.5.4]
 
 ```plain